api/pyproject.toml
[tool.poetry]
name = "radar-api"
version = "0.1.0"
description = ""
authors = ["forefy <tomer@auditware.io>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
django-jazzmin = "^3.0.0"
djangorestframework = "^3.15.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
